남의글 지우기 exception 구현

+++ corona/src/main/java/com/gura/corona/cafe/service/CafeServiceImpl.java
@@ -1,5 +1,6 @@
 package com.gura.corona.cafe.service;
 
+import java.net.URLEncoder;
 import java.util.List;
 
 import javax.servlet.http.HttpServletRequest;
@@ -9,17 +10,89 @@ import org.springframework.stereotype.Service;
 
 import com.gura.corona.cafe.dao.CafeDao;
 import com.gura.corona.cafe.dto.CafeDto;
+import com.gura.corona.exception.NotDeleteException;
 
 @Service
 public class CafeServiceImpl implements CafeService{
 	@Autowired
 	private CafeDao cafedao;
-
+	
+	
+	//한 페이지에 나타낼 row 의 갯수
+	final int PAGE_ROW_COUNT=5;
+	//하단 디스플레이 페이지 갯수
+	final int PAGE_DISPLAY_COUNT=5;
+	
 	@Override
 	public void getList(HttpServletRequest request) {
 		
-		List<CafeDto> list=cafedao.getList();
-		request.setAttribute("list", list);
+		//보여줄 페이지의 번호
+				int pageNum=1;
+				//보여줄 페이지의 번호가 파라미터로 전달되는지 읽어와 본다.	
+				String strPageNum=request.getParameter("pageNum");
+				if(strPageNum != null){//페이지 번호가 파라미터로 넘어온다면
+					//페이지 번호를 설정한다.
+					pageNum=Integer.parseInt(strPageNum);
+				}
+				//보여줄 페이지 데이터의 시작 ResultSet row 번호
+				int startRowNum=1+(pageNum-1)*PAGE_ROW_COUNT;
+				//보여줄 페이지 데이터의 끝 ResultSet row 번호
+				int endRowNum=pageNum*PAGE_ROW_COUNT;
+				/*
+					검색 키워드에 관련된 처리 
+				*/
+				String keyword=request.getParameter("keyword"); //검색 키워드
+				String condition=request.getParameter("condition"); //검색 조건
+				if(keyword==null){//전달된 키워드가 없다면 
+					keyword=""; //빈 문자열을 넣어준다. 
+					condition="";
+				}
+				//인코딩된 키워드를 미리 만들어 둔다. 
+				String encodedK=URLEncoder.encode(keyword);
+				
+				//검색 키워드와 startRowNum, endRowNum 을 담을 FileDto 객체 생성
+				CafeDto dto=new CafeDto();
+				dto.setStartRowNum(startRowNum);
+				dto.setEndRowNum(endRowNum);
+				
+				if(!keyword.equals("")){ //만일 키워드가 넘어온다면 
+					if(condition.equals("title_content")){
+						//검색 키워드를 FileDto 객체의 필드에 담는다. 
+						dto.setTitle(keyword);
+						dto.setContent(keyword);	
+					}else if(condition.equals("title")){
+						dto.setTitle(keyword);
+					}else if(condition.equals("writer")){
+						dto.setWriter(keyword);
+					}
+				}
+				//파일 목록 얻어오기
+				List<CafeDto> list=cafedao.getList(dto);
+				//전체 row 의 갯수 
+				int totalRow=cafedao.getCount(dto);
+				
+				//전체 페이지의 갯수 구하기
+				int totalPageCount=
+						(int)Math.ceil(totalRow/(double)PAGE_ROW_COUNT);
+				//시작 페이지 번호
+				int startPageNum=
+					1+((pageNum-1)/PAGE_DISPLAY_COUNT)*PAGE_DISPLAY_COUNT;
+				//끝 페이지 번호
+				int endPageNum=startPageNum+PAGE_DISPLAY_COUNT-1;
+				//끝 페이지 번호가 잘못된 값이라면 
+				if(totalPageCount < endPageNum){
+					endPageNum=totalPageCount; //보정해준다. 
+				}
+				
+				//EL 에서 사용할 값을 미리 request 에 담아두기
+				request.setAttribute("list", list);
+				request.setAttribute("startPageNum", startPageNum);
+				request.setAttribute("endPageNum", endPageNum);
+				request.setAttribute("pageNum", pageNum);
+				request.setAttribute("totalPageCount", totalPageCount);
+				request.setAttribute("condition", condition);
+				request.setAttribute("keyword", keyword);
+				request.setAttribute("encodedK", encodedK);	
 	}
 
 	@Override
@@ -38,7 +111,13 @@ public class CafeServiceImpl implements CafeService{
 	@Override
 	public void delete(HttpServletRequest request) {
 		int num=Integer.parseInt(request.getParameter("num"));
-		
+		//삭제할 글 정보를 얻어온다.
+		CafeDto cafeDto=cafedao.getData(num);
+		//세션에 저장된 아이디를 읽어온다(로그인된 아이디)
+		String id=(String)request.getSession().getAttribute("id");
+		if(!id.equals(cafeDto.getWriter())) {
+			throw new NotDeleteException("남의글 지우기없기");
+		}
 		cafedao.delete(num);
 		
 	}
